+++ src/main/kotlin/com/zepben/evolve/database/sqlite/cim/CIMDatabaseReader.kt
@@ -78,15 +78,13 @@ abstract class CimDatabaseReader(
 
     private fun preLoad(): Boolean =
         try {
-            connection.createStatement().use { statement ->
-                val version = tableVersion.getVersion(statement)
-                if (version == supportedVersion) {
-                    logger.info("Loading from database version v$version")
-                    true
-                } else {
-                    logger.error(formatVersionError(version))
-                    false
-                }
+            val version = tableVersion.getVersion(connection)
+            if (version == supportedVersion) {
+                logger.info("Loading from database version v$version")
+                true
+            } else {
+                logger.error(formatVersionError(version))
+                false
             }
         } catch (e: Exception) {
             logger.error("Failed to connect to the database for reading: " + e.message, e)
@@ -99,7 +97,7 @@ abstract class CimDatabaseReader(
 
     private fun formatVersionError(version: Int?): String =
         when {
-            version == null -> "Failed to read the version number form the selected database. Are you sure it is a EWB database?"
+            version == null -> "Failed to read the version number from the selected database. Are you sure it is a EWB database?"
             version < supportedVersion -> unexpectedVersion(version, "Consider using the UpgradeRunner if you wish to support this database.")
             else -> unexpectedVersion(version, "You need to use a newer version of the SDK to load this database.")
         }