+++ src/main/kotlin/com/zepben/evolve/database/sqlite/cim/upgrade/UpgradeRunner.kt
@@ -10,9 +10,9 @@ package com.zepben.evolve.database.sqlite.cim.upgrade
 
 import com.zepben.evolve.database.paths.DatabaseType
 import com.zepben.evolve.database.sqlite.cim.tables.tableCimVersion
-import com.zepben.evolve.database.sqlite.cim.upgrade.changesets.*
-import com.zepben.evolve.database.sqlite.common.TableVersion
 import com.zepben.evolve.database.sqlite.extensions.configureBatch
+import com.zepben.evolve.database.sqlite.common.TableVersion
+import com.zepben.evolve.database.sqlite.cim.upgrade.changesets.*
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
 import java.io.IOException
@@ -99,7 +99,7 @@ class UpgradeRunner @JvmOverloads constructor(
         backupRequired: Boolean
     ): UpgradeState =
         connection.createStatement().use { statement ->
-            tryRunUpgrade(tableVersion.getVersion(statement), changeSets.last().number) { databaseVersion ->
+            tryRunUpgrade(tableVersion.getVersion(connection), changeSets.last().number) { databaseVersion ->
                 logger.info("Upgrading database '$databaseFile' from v$databaseVersion to v${changeSets.last().number}")
 
                 if (backupRequired)