+++ src/test/kotlin/com/zepben/evolve/database/sqlite/cim/upgrade/changesets/customer/ChangeSet50CustomerValidator.kt
@@ -9,8 +9,8 @@
 package com.zepben.evolve.database.sqlite.cim.upgrade.changesets.customer
 
 import com.zepben.evolve.database.paths.DatabaseType
-import com.zepben.evolve.database.sqlite.cim.upgrade.changesets.ChangeSet50Helpers
 import com.zepben.evolve.database.sqlite.cim.upgrade.changesets.ChangeSetValidator
+import com.zepben.evolve.database.sqlite.cim.upgrade.changesets.ChangeSet50Helpers
 import java.sql.Statement
 
 object ChangeSet50CustomerValidator : ChangeSetValidator {